Source: https://github.com/msupply-foundation/open-msupply/commit/4e98db6d8c64065e00bfd7ed085d4f1d0213fb23

+++ server/service/src/sync/translations/name_insurance_join.rs
@@ -55,7 +55,7 @@ pub struct LegacyNameInsuranceJoinRow {
     pub ID: String,
     pub nameID: String,
     pub insuranceProviderID: String,
-    pub discountRate: i32,
+    pub discountRate: f64,
     #[serde(deserialize_with = "empty_str_as_option_string")]
     pub enteredByID: Option<String>,
     pub expiryDate: NaiveDate,
